added confirmation for delete notes in details view

+++ note-taker/src/app/components/note-detail/note-detail.component.ts
@@ -2,16 +2,19 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NoteService } from '../../services/note.service';
 import { PrimaryActionDirective } from '../../directives/primary-action.directive';
+import { ModalDialogComponent } from '../../modal-dialog/modal-dialog.component';
+import { UserAction } from '../../models/model';
 
 @Component({
   selector: 'app-note-detail',
   standalone: true,
-  imports: [PrimaryActionDirective],
+  imports: [PrimaryActionDirective, ModalDialogComponent],
   templateUrl: `./note-detail.component.html`,
   styleUrl: './note-detail.component.scss',
 })
 export class NoteDetailComponent implements OnInit {
   note: any;
+  confirmDeleteModalVisible = false;
 
   constructor(private route: ActivatedRoute, private noteService: NoteService, private router: Router) { }
 
@@ -29,9 +32,17 @@ export class NoteDetailComponent implements OnInit {
   }
 
   deleteNote() {
-    const id = this.route.snapshot.paramMap.get('id');
-    this.noteService.deleteNote(Number(id)).subscribe(() => {
-      this.router.navigate(['/']);
-    });
+    this.confirmDeleteModalVisible = true;
+  }
+
+  onUserAction(event: UserAction) {
+    if (event == UserAction.Primary) {
+      const id = this.route.snapshot.paramMap.get('id');
+      this.noteService.deleteNote(Number(id)).subscribe(() => {
+        this.router.navigate(['/notes'])
+      });
+    } else {
+      this.confirmDeleteModalVisible = false;
+    }
   }
 }
